requirements.txt
python-dotenv
langchain
langchain-groq
langchain-mcp-adapters
mcp
langgraph
duckduckgo_search
ddgs
